dashboard/requirements.txt
fastapi
pyserial
uvicorn[standard]
